requirements.txt
streamlit
playwright
moviepy
openai-whisper
openai
edge-tts
requests
python-dotenv
yt-dlp
